+++ rw-csv/rw.h
@@ -4,7 +4,7 @@
 
 #include "../point/point.h"
 
-void read_csv (int, int, const char *, int **);
+void read_csv (int, int, const char *, double **);
 
 void create_marks_csv(point *, int, const char *);
